Updated tags taxonomy features export
The hierarchy value depends on the state of the content and not on the configuration for the taxonomy itself. The content has changed and thus the value should be updated as well.

This has little to do with the purpose of this branch but we need to update the feature so we fix this as well.